backend/requirements.txt
# Vivian Cosmetic Shop - Backend Dependencies
# Python 3.9+

# Flask Framework
Flask==3.0.0
flask-cors==4.0.0

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyMySQL==1.1.0
cryptography==41.0.7

# Authentication
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1

# Environment
python-dotenv==1.0.0

# Utilities
email-validator==2.1.0

# Development
pytest==7.4.3
pytest-flask==1.3.0
